Encapsulación suite-test en su propio componente
